+++ frontend/src/scenes/events/UsageDisabledWarning.tsx
@@ -1,0 +1,21 @@
+import { Alert } from 'antd'
+import React from 'react'
+
+export function UsageDisabledWarning({ tab }: { tab: string }): JSX.Element {
+    return (
+        <Alert
+            type="info"
+            showIcon
+            message={`${tab} is not enabled for your instance.`}
+            description={
+                <>
+                    You will still see the list of events and properties, but usage information will be unavailable. If
+                    you want to enable event usage please set the follow environment variable:{' '}
+                    <pre style={{ display: 'inline' }}>ASYNC_EVENT_PROPERTY_USAGE=1</pre>. Please note, enabling this
+                    environment variable <b>may increase load considerably in your infrastructure</b>, particularly if
+                    you have a large volume of events.
+                </>
+            }
+        />
+    )
+}
