+++ frontend/src/scenes/events/VolumeTable.scss
